Automate multiple feed discovery and generation
Feeds are now configured by "Atom seeds", Markdown files with a YAML
frontmatter, which exist a `./feeds/` directory. Their filepath,
excluding the `./feeds/` prefix, should mirror existing directories with
Markdown entries within them, and will be converted to a full Atom feed
of that directory. This avoids extra config.

A new `./_ignore` file is used and contains patterns of filepaths to
avoid converting, like a `.gitignore`.